Fix: stop user (floating/dockable) windows from being moved by setWindow()
Otherwise it splits the `TConsole` from it's parent `TDockWidget` and it
cannot be put back; they also share the same name internally which makes
things confusing if they get divorced from each other.

The same could also apply if the map is being displayed in it's own
floating/dockable widget, so detect, abort and report if that is the case,
or if the window that that is in is given as the destination for any
element (otherwise the message that the "element 'mapper' not found" is
confusingly misleading)!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>