Revise: record which command line is being used for each profile
So that we can return the focus to it when we switch between profiles.

This is an amalgamation with the code from #6232 and recent changes in the
main `development` branch. However because the development of this PR
has taken so long it has been difficult to clean it up whilst also merging
in the other changes so I have taken the step of saving copies of all the
files that my working code ends up as and then injected them into the last
published commit. This might seem messy but it was the only way I could see
to keep the bits I wanted but not the bits I didn't.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>