Fix: remove many setFocusPolicy(...) statements that are probably unneeded
I found out that setting a focusPolicy on a widget that itself has a
focusProxy (so that it sends the focus on it to that proxy) has the effect
of applying the policy to the *proxy*. Unfortunately that meant that
somehow with the previous commits to this PR, it was possible for the
main command line to end up with a "NoFocus" policy applied to it with the
result that after creating or enabling other `TCommandLine` instances for
a profile it became *impossible* to click on the main command line so that
it got the focus and was able to receive keyboard input.

As the default focusPolicy for *most* widgets is `NoFocus` anyhow I felt it
was probably safe to remove any policy settings, particular to that type,
so it would not get sent/set accidentally on the main command line when it
was the focusProxy for those widgets. The cases where it might be required,
i.e. any `TCommandLine`, are set to `StrongFocus` in the constructor.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>