Fix: ensure 3D push-button is updated if mode is changed by `setConfig(...)`
Before this PR changing the setting did not update the button with the
result that if it was changed by the Lua API once the next click of the
button would not produce any change as it changed to match the actual
state.

Also adds a getter `(bool) isIn3DMode()` to the `dlgMapper` class so that
other parts of the application can find out the current setting without
poking around too deeply in that class.

This will close an issue for this particular item that I raised in
our Discord **mudlet-devlopment** channel:
https://canary.discord.com/channels/283581582550237184/283582439002210305/1078169157155160075

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>