Fix: get Mudlet starting after installation on Windows
This includes additional diagnostic code that will put a `[ WARN ] -`
message up in the main console of the first profile opened after Mudlet
is started with a command line argument that is not recognised.

This should close #6500 as it appears that that is being caused by a
command line option from the installer and since #6470 Mudlet would abort
and show the `-h`/`--help` output should it get a command line argument
that it doesn't recognise.

This is an alternative to #6582 which instead removes the non-zero exit
code and reports the parsing error message to the OS command line rather
than to the game which means it may not be visible to the user.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>