fix: 3d mapper can now utilize custom speedwalks (#6534)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Double click on 2D map will check if you have a custom speedwalk, single
click a room on 3D map which also does speedwalks did not get same
change. The `speedWalkDir` and `speedWalkPath` were set in both 2D and
3D scenarios, but `speedWalkFrom` and `speedWalkTo` was only set with
the 2D.

Looked at changes in #4156. Took stuff from T2DMap.cpp:545
T2DMap::initiateSpeedWalk and pasted it in with the click event in the
glwidget.cpp:2100 GLWidget::mousePressEvent where the existing speedwalk
stuff is.

#### Motivation for adding to Mudlet
I have a custom `doSpeedWalk()` function which has some logic to allow
creators to teleport directly to the room instead, in cases when there
is not a path linking to the room. It was confused at `speedWalkFrom`
and `speedWalkTo` not being updated, or being nil, depending on if I had
done a 2d speedwalk beforehand. It starts out though by relying on those
2 variables having been set, starts with `if(getPath(speedWalkFrom,
speedWalkTo))`. So it would end up doing wrong directions or an error
due to nil. This mapper that I'm using is set to download on
Cheeseworld, but you could try this on any mapper that has custom
speedwalk stuff in it.

#### Other info (issues closed, discussion etc)
Summary above is basically what the discussion was, here is a link
though.

https://discord.com/channels/283581582550237184/283582439002210305/1070745380985241620

---------

Co-authored-by: Stephen Lyons <slysven@virginmedia.com>