fix: launch Mudlet anyway when given unknown command line options (#6582)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Refrain from shutting down when an unrecognized command line option was
given. Display a warning, but do not shut down as if user had started
the program from command line and requested help via command line.
#### Motivation for adding to Mudlet
In #6470 the scope got expanded from just handling `--profile` better to
also making unknown command line options act as as `--help`. It turns
out that the installer launches the program with `--squirrrel-firstrun`
at end of install to facilitate "Welcome" and "What's new" type of
stuff. So this meant that the final step of installation would mess up.
We thought it was not launching either due to update code or installer
code, but it was in fact launching and then immediately shutting down.
Because it wasn't launched from command line, we were not seeing the
message sent to standard output. This change simply undoes that part.
Having a handler for that specific option would be good.
#### Other info (issues closed, discussion etc)
FIxes #6500.