fix recognizing that downloadable GUI is already installed (#6584)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Prior to #6541 the Client.GUI handler would recognize package is already
installed and then return. Then the returns were taken out in order to
continue on to the variable setting part a few lines later, so that
users could see the data via the gmcp variable. But, just after that
return and prior to the variable setting, there is some code for when
package is NOT already installed. Long story short, if package is
already installed then now you get a message saying package is already
installed, and then it proceeds to reinstall package.

To change that, I put the next chunk of code for package installation
inside of an `else{ }`.
#### Motivation for adding to Mudlet

#### Other info (issues closed, discussion etc)
Fixes #6580 

Clessidra sends the Client.GUI message with the version number and URL
instead of the JSON format. So if you test out the variable saving part
then use StickMUD instead. (Both formats are great, just the other one
won't save it to a variable where scripts can see it)