Improve: add ancestors functions
Provides two Lua functions:

* `isAncestorsActive(itemID, "type") - returns a boolean which is `true` if
all (if any) of the parents of the given item are set as active and `false`
if any are not. It makes use of an internal template function that already
exists but makes it available to the Lua sub-system. If the item does not
have any ancestors this function will return `true`. This will be enough to
close #6724.

* `ancestors(itemID, "type") - returns a table containing a sub-table for
each successively distance ancestor (if any) of the given item; within
each sub-table are details of the ancestor, specifically:
  * its ID as a number
  * its name as a string
  * whether it is active as a boolean
  * its "node" (type), one of "item", "group" (folder) or "package" (module)

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>