Infrastructure: try to fix failure to update dblsql feed with PTB builds (#6605)
This (the `link-ptbs-todblsqqd` GHA job) has been failing since
2023/01/02 and it seems that the version of Luarocks expects a later
`libc` than is present in the Ubuntu version "Focal LTS" 20.04 that we
are using. This updates the Linux OS used for that GH Action to the
Ubuntu *latest* which is now 22.x. Hopefully this will restore
operation.

Also update the `tag-pull-request` as it also was using the same older
Ubuntu version.

Furthermore this PR also adds two other recently added GHA job files to
the files shown in Qt Creator - which makes it easier to get to them in
that environment.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>