Fix: correct search for Lua interpreter for qmake builds
Discord user **#Caith#8202** discovered that building Mudlet on aarch64 i.e.
the Arm64 platform on Fedora fails (see the Mudlet **#help** channel on
Discord starting 2023/03/06 03:55 UTC with the text):
> lua-yajl is the bane of my existence.  edited the hell out of the cmake
scripts and mudlet.pro to finally get it to start using lua 5.1 instead of
5.4, then had to rig up some hunspell junk, but no matter what i do, qmake
won't cooperate with the lua-yajl install.
>
> compiles fine on ubuntu and arch, so i'm bewildered.

One detail that came to light is that on that system the 5.1 version of the
Lua interpreter is called `lua-5.1` but that is not one of the names that
`./translations/translated/updateqm.pri` tries so it fails to
generate/update the binary translations for inclusion in the project.

This PR should extend the tries to a fourth one that will handle that form,
it also restructures the code slightly as it looked to me as the Windows
OS handling was off.

The corresponding CMake build file has also been updated to handle the same
four variants - prior to that it only tried two which might not have worked
in all cases.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>