Fix: restore generation of `sysWindowsResizeEvent` when borders are changed (#6620)
This should close #6614.

Whilst looking through the code-base I was confused to see that the
members `(int) mMainFrame{Top|Bottom|Right|Left}{Height|Width}` in the
`TConsole` class were directly related to the `(int)
mBorder{Top|Bottom|Left|Right}{Height|Width}` members. Also they are
often manipulated as a set. I felt that it was constructive to refactor
them into a common `std::tuple<int, int, int, int> mBorders`.

This refactoring meant it was also possible to detect when the borders
were changed (in the new `(void) Host::setBorders(const std::tuple<int,
int, int, int> borders)`) and to raise the missing
`sysWindowResizeEvent` as needed.

Further research revealed that the Qt `QMargin` class was perfect for
this task and easier to read than a `std::tuple<int, int. int, int>` as each
of the four members have a distinguishable name instead of having
to use an index number. Therefore the code was rewritten to use that
class instead.
 
Signed-off-by: Stephen Lyons <slysven@virginmedia.com>
Co-authored-by: Kebap <kebap_spam@gmx.net>