Infrastructure: rationalise strings in `TLuaInterpreter` class
In particular try and use ID in a uniform manner when referring to an
integer number that uniquely identifies an instance of something like
a room or an area in the map.

It became clear that there are a significant number of duplicate strings
involving ID and it was possible to reduce the number of duplicate
`qsl(`...`)` - which are our wrapper of `QStringLiteral` - by replacing
each with a QString constant initialised at compile time. This is a good
thing as it is bad to have multiple copies of the same string as one of
these as each one is stored separately in the (Read-Only) data part of the
executable.

This should close #2605.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>