Fix: Make Geyser less picky about object names (#6628)
#### Brief overview of PR changes/additions
Someone was having an issue in Discord today and it turned out their
object name had "Class" in it, which was causing it to error. This code
is meant to filter out the geyser class objects, but was a bit too
broad. This will now only filter things out if their name is
`<something>Class`. All the Geyser class names match this pattern
already but it should reduce confusing errors in future.
#### Motivation for adding to Mudlet
Less unexpected and confusing errors is better.
#### Other info (issues closed, discussion etc)