Infrastructure: minor tweaks
In `main.cpp`:
* remove unused `(TConsole*) spDebugConsole`

* remove Q_OS_UNIX only setting of `QApplication` flag
`Qt::AA_EnableHighDpiScaling` as that is set in the `mudlet` cosntructor
for ALL OSes.

Refactor the code that selects the background image for the Splash-screen
to a separate static method in the `mudlet` class as it is used both in
main.cpp AND in the `dlgAboutDialog` class. This thus is a DRY step.

Also, if Mudlet is being built with a constant splash-screen do not include
the special ones for the PTB and DEV builds as they are not then reguired.
They have been moved to a separate resource file which is only included
if required.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>