Improve: remember 2D mapper zoom amounts between sessions (#6615)
This is intended to close #2388.

This enables the 2D (only) map zoom amounts for each map area to be save
independently and restored when switching between areas in the mapper.
It also saves the data between session.

It extends the existing `setMapZoom(...)` API to take an optional second
argument to specify the map area ID to set the zoom (which is a floating
point number) for any existing area, not just the current one.

It also adds a `getMapZoom(...)` function that, without any arguments,
returns the currently used 2D map zoom value for the area currently
being shown in the 2D mapper. If an area ID is provided it instead
returns the value that was last used for that area - or the default
value that is used initially on starting the profile or for an area that
has not been viewed before.

Importantly when switching between the areas in the 2D mapper the values
are retained and applied so that one area can be zoomed in and another
zoomed out and switching from the first to the second and back to the
first means that the zoom level used in the first is reused when it is
returned to. Deleting an area will forget the stored zoom level so if it
is reused it starts from scratch.

Code to save the zoom level for each area has also been implemented
within the C++ core. It saves it in the Area User Data for current map
formats (but removes it on loading so the user never sees it there)
under a `system.fallback_map2DZoom` key but will save it directly in the
binary data (which is more efficient) in the next format version
whenever it is enabled.

A new Mudlet event, which has been called `sysMapAreaChanged` has
been added with two additional arguments being the area ID changed to
followed by the one that it was changed from. I originally thought I
would need it to handle saving the zoom level for each area via the Lua
system but I found that that was not practicable.

Also, in refactoring `T2DMap::paintEvent(...)` I:
Removed/combined some locals:
* `(TArea*) playerArea` and `pPlayerArea` and `pArea`==> `pDrawnArea`
* `(TRoom*) playerRoom` ==> `pPlayerRoom`

Remove unneeded (refactored out):
* `(qreal) ox`
* `(qreal) oy`

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>