Fix: prevent right clicking in error console moving focus to main window
This should close #6638 as it prevents that and the Central Debug Console
windows from performing the step in the TTextEdit::mouseReleaseEvent(...)
that was redirecting the focus to the profile's currently used command-
line. It also fixes a related matter in that those windows were also
possibly raising bogus `sysWindowMousePressEvent` and
`sysWindowMouseReleaseEvent` events that a user's GUI system would not
want.

For some reason Qt Creator has also taken upon itself to fix up some
spacing issues which are in accordance to our preferences so I have left
them in the commit.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>