Fix: prevent user map menus being deleted prematurely
This will close #6642. It was happening because the (sub) menus were being
set to automatically delete when they are closed - which happens when the
user selects an item on them, or, after a delay, if the pointer moves off
of them. In the latter case this is what was causing the problem as the
parent menu (or even a further up ancestor) does not necessarily go away
and it is expected the sub-menu should persist. The cure is to remove the
automatic delete flag from the sub-menus. However in itself this means that
they would persist in memory until the widget that owns them is destroyed,
as this is the `T2DMap` instance that could lead to 1000s of them being
leaked in this manner. The solution is to tell them that their parent is
the top-level (context) menu so that when it goes they do as well and not
before!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>