fix F6 keybinding (#6670)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
The accessibility code has special behavior for F6 that is optional, but
if the option is off or assigned to something other than F6, it gets
interrupted and doesn't run keybinds set by user. Also I'll make it more
strict when it IS enabled, and make it so that modifiers like Ctrl-F6
are not treated as F6.

#### Motivation for adding to Mudlet
Fame and accolades

#### Other info (issues closed, discussion etc)
Fixes #6666

(cherry picked from commit 703e2f9825ee8257f8a7034575950e12d2598c61)