Fix: needed/unneeded context for lambdas
Clazy was picking up "connect-3arg-lambda" warnings, explained as item 3
here: https://www.kdab.com/nailing-13-signal-slot-mistakes-clazy-1-3 .

It seemed to fix crashes I was getting when I was trying to change the
password secure storage setting, see:
https://github.com/Mudlet/Mudlet/issues/6671#issuecomment-1483211392
however those were not happening every time - but that could be down to
which threads the two ends of a connect was happening on, which the extra
argument helps to determine AFAICT.

Also:
* remove some redundant `QObject` prepended to `connect(...)` calls
* rename some arguments passed into some lambda so that they do not shadow
other variables in the parent methods.
* identify `dlgProfilePreferences::hidePasswordMigrationLabel()` as a Qt
SLOT, renamed to `slot_hidePasswordMigrationLabel()`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>