Fix: profile tabs bolded by activity staying bold after switching to them (#6690)
<!-- Keep the title short & concise so anyone non-technical can
understand it,
     the title appears in PTB changelogs -->
#### Brief overview of PR changes/additions
Profile tabs become bold when there is activity and not looking at that
tab. Switching to that tab is supposed to clear the bold but currently
it is not. This is meant to fix that.
#### Motivation for adding to Mudlet
Fix #6672
#### Other info (issues closed, discussion etc)
The #6267 had a comment `Moved as much as possible to activateProfile`
and then found this chunk that got lost. Just copied and pasted from
there.

(cherry picked from commit 057e02e7a53dd4bc9dbd809d3041735406a3f158)