Fix: do not store/load profile name, url or port in game save files
This is counterproductive as the data is supposed to be stored in the base
directory of the profile and changes made in the "Connect profile" dialog
get overridden by the game save.

I nearly got banned for multi-playing on a Server because I was working on
a separate issue related to multiple auto loading profiles and two of the
profiles I normally used had been edited to set the URL to `localhost` but
when the profiles were autoloaded they used the stored settings and
promptly BOTH connected to the Game server - waking up its anti-
multi-playing detection!

As the profile name is not now being stored within the profile save file
there is no need to backup and then restore it when installing a package,
so code to do that and the same for the character name and login password
is also removed from `Host::installPackage(...)`.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>