Add: allow historical maps to be selected on opening profile
This should close #656!

This adds another `QComboBox` selector under the game save selector in the
second (options) tab on the "Connect profile" dialogue to allow any map
(Mudlet `.dat` binary, `.json` or `.xml`) file stored in the "normal place"
to be chosen and loaded instead of the default "last" one. Like the game
save selection an icon is pre-pended to the "autosave" one, in addition a
(icon theme dependent) icon is pre-pended for `.json` and `.xml` files so
they can be distinguished from the "normal" binary one. Like the game save
list the entries are sorted into descending date and time order.

Also, renames:
* `(QComboBox*) dlgConnectProfiles::profile_history`
                                              ==> `comboBox_profileHistory`

For optimisation reasons the `QRegularExpression`s used to detect the
details of normal game and map save files have been made into
`static const`s so that they can be optimised and created just once, rather
than every time a different profile in the "Connect profile" dialog is
opened.

It became clear that some lines for odd files could be really long and
something to adjust the widths of some widgets in the "Connect profile"
dialog would be beneficial. I have added code that tries to do that, it
should ensure that there is enough space on the left to show three columns
of profile icons and enough on the right to show the full length of the
longest file selected in the two combo boxes for the game save and map
files for the currently selected profile.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>