Fix: remove DejaVu fonts from being specified
I spotted on the MUME Discord that they were wondering about why mentions
of Deja Vu fonts were occuring in some files (HTML?) that Mudlet was using/
producing on Windows machines. I remembered that we shipped a few fonts
but noticed that we bundle the earlier "Bitstream Vera" fonts (which were
one of the contributors to the DejaVu ones) and that we do NOT include the
latter.  As such it isn't sensible for us to specify (in the places that
we do) the DejaVu font when the bundled Bitstream one is suitable. In
checking for places where we do mention a font I spotted that there were
some probably bogus ones in some Qt Designer forms/dialogs for specific
widgets that should probably be removed so that a parent one for the whole
dialogue (if provided) was used - leaving a size parameter where that was
wanted.

Whilst working on this I spotted some issues with the way the size of the
font used to draw the network and system latencies was calculated was done.
The "dummy" version of the text used in a three size test was not denoted
as translatable texts so the width that the process would decide upon would
not match the translated text used in non-en locales and also it only tried
three different sizes - which might not be optimum for some locales. I
refactored and redesigned this bit of code to use a loop and try more sizes
and use the translated form of the text!

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>