Fix: allow search highlighting to be undone
This adds a clear button to the Search widget on the Main Console's command
line which also clears all the (yellow on black) highlights used during
a search - and entering a different term also clears the highlights ready
for the next term.

It also should accommodate when the underlying `TBuffer` is shrunk because
it has exceeded the maximum number of lines - which the prior code didn't.

This should close #3759.

There is future scope to speed up the clearing process - I intend to
convert the `(QList<bool>) TBuffer::promptBuffer` to a
`(QList<LineFlags>) TBuffer::lineFlags` construct where the flag that a
line is a "prompt" one is just one bit - and the fact that the line
contains something that matches the current search term is another. So that
when iterating through the whole buffer to clear the new
`TChar::AttributeFlag::SearchFind` flag for those characters that are high-
lighted Mudlet instead only has to go over those lines with the line flag
set. {Such a line flag structure will also assist in my evil plans to
implement flashing text so it is possible to work out which lines contain
text that has to be redrawn every so often to make it flash!}

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>