Fix: instability when auto-loading multiple profiles (#6712)
This tries to avoid multiple single shot timers from being created to
move the focus on the appropriate command line of a particular profile.
However the actual fix in the situation of autoloading profiles is to
remove the code that generated a call to switch to a profile in the
`focusInEvent` for the `TCommandLine` class - in combination these two
changes should prevent the unstable, and continuous, oscillation between
two profiles I saw before making these changes. This should close #6669.

Furthermore during the creation of the first profile something was not
being completed that meant the profile was not capable of being hidden
when a second profile was loaded afterwards. This can be corrected by
allowing pending events (such as pending "zero-timers") to complete,
this may improve #6392...

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>