Fix: ensure map loading does not take place until Lua sub-system is started
The `TMap::audit()` function - which produces the `[ OK ]`  message for a
map loading operation makes use of an external Lua function, so it cannot
be used until the Lua sub-system has been set up. Doing the map loading
from a zero-timer at the end of the `Host` constructor is too early to allow
this to work properly - so it needs to be moved to something that can be
called after then.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>