Fix: ensure map loading occurs at correct point in profile initialisation
For some time now (years) the loading of the default (last saved) map file
was done via a "zero-timer" started at the end of the `Host` constructor.
However we are now using several such zero-timers and the Qt documentation
does point out that the order of triggering for more that one is arbitrary.
This PR changes the point at which the map is loaded to a precise point in
the profile initialisation sequence to be just after the Lua sub-system is
setup but before the connection to the server (if to be done) will occur.

Signed-off-by: Stephen Lyons <slysven@virginmedia.com>